Move on_encrypt function to its own file
