+++ verification/cbmc/proofs/default_cmm_generate_enc_materials/default_cmm_generate_enc_materials_harness.c
@@ -27,40 +27,7 @@ int on_encrypt(
     struct aws_array_list *keyring_trace,
     struct aws_array_list *edks,
     const struct aws_hash_table *enc_ctx,
-    enum aws_cryptosdk_alg_id alg) {
-    /* Check validity for all inputs to avoid memory-safety violations. */
-    assert(aws_cryptosdk_keyring_is_valid(keyring));
-    assert(aws_allocator_is_valid(request_alloc));
-    assert(aws_byte_buf_is_valid(unencrypted_data_key));
-    assert(aws_cryptosdk_keyring_trace_is_valid(keyring_trace));
-    assert(aws_cryptosdk_edk_list_is_valid(edks));
-    assert(aws_cryptosdk_edk_list_elements_are_valid(edks));
-    assert((enc_ctx == NULL) || aws_hash_table_is_valid(enc_ctx));
-
-    if (unencrypted_data_key->buffer == NULL) {
-        const struct aws_cryptosdk_alg_properties *props = aws_cryptosdk_alg_props(alg);
-        if (props != NULL) {
-            __CPROVER_assume(aws_byte_buf_is_bounded(unencrypted_data_key, props->data_key_len));
-            ensure_byte_buf_has_allocated_buffer_member(unencrypted_data_key);
-            /*
-             * This satisfies the post-condition that if this keyring
-             * generated data key, it must be the right length.
-             * The nondeterminism increases coverage.
-             */
-            if (nondet_bool()) unencrypted_data_key->len = props->data_key_len;
-            unencrypted_data_key->allocator = request_alloc;
-            __CPROVER_assume(aws_byte_buf_is_valid(unencrypted_data_key));
-        }
-    } else {
-        /*
-         * If the buffer is not NULL, the byte buffer must not have been modified;
-         * however, this allocation modifies the buffer and increases coverage.
-         */
-        if (nondet_bool()) ensure_byte_buf_has_allocated_buffer_member(unencrypted_data_key);
-    }
-    int ret;
-    return ret;
-}
+    enum aws_cryptosdk_alg_id alg);
 
 void default_cmm_generate_enc_materials_harness() {
     /* Nondet input required to init cmm */
